cflat_runtime: recover request structure (79.49% -> 89.80%)
- Rewrite request() with real CObject/CFunc member access
- m_codeIndex bitfield: short:12/int:20 mixed containers reproduce the
  target's lha/rlwimi/sth + lwz/rlwimi/stw write pattern exactly
- getTopBit: loop form with early returns; inlined call yields the
  target's beq/b branch web
- pop() moved above request so it inlines (union temp produces the
  stack-home spill seen in the target)
- copy loops in difference-condition form to avoid the -O4,p unroller
- bitfield m_callFlag read/write instead of manual masks

+++ include/ffcc/cflat_runtime.h
@@ -51,7 +51,7 @@ public:
 		union {
 			unsigned int m_codePos; // 0x1C
 			struct {
-				int m_codeFunc : 12;
+				short m_codeFunc : 12;
 				int m_codeOffset : 20;
 			} m_codeIndex;
 		};
